chore: moved histories to separate headers
Bench: 4959499
