+++ src/core/board/board.h
@@ -6,7 +6,7 @@
 
 #include "attacks/attacks.h"
 #include "defs.h"
-#include "search/history.h"
+#include "search/history/history.h"
 #include "move.h"
 #include "evaluation/nnue/nnue.h"
 #include "types.h"
